functions/parse_email_function/temp/requirements.txt
azure-functions==1.20.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
filelock==3.15.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
requests==2.32.3
requests-file==2.1.0
sniffio==1.3.1
soupsieve==2.5
tldextract==5.1.2
urllib3==2.2.2
anyio==4.4.0